Added functionality to check if wordpress already exists in the user's directory.
This wordpress check is used when rendering the tools template, so as to give the user an option to install wordpress if it is not already installed.
If wordpress is not installed, the user has the option to click install. This creates a an AJAX request to a newly created API endpoint "/wordpressinstall", which is used to install wordpress for a particular user.
After installing, a link for the user's wordpress site is revealed.