Changes the help email so that it sends a single email to all SysAdmins
The help email is now sent to the netsoc email address, CC'ing all of the SysAdmins
and the user that requested help. This means that all of the SysAdmins can see when
someone has responded to an email as well as giving everyone the full context of everything
that has been done for the user up until that point.